requirements.txt
Flask-WTF
redis
python-dotenv
requests